projeler/haber-bot/Ajan-bot/requirements.txt
# Ajan-Bot — Fuar Lead Generator
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0

# Database
# SQLite is built-in

# Telegram
python-telegram-bot>=20.7

# Utils
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
tqdm>=4.66.0

# Email validation
email-validator>=2.1.0
